+++ src/behaviours/__tests__/index.test.tsx
@@ -3,46 +3,11 @@ import { useState } from "react"
 import "../../../jest.setup"
 import { motion } from "../../"
 import { motionValue, MotionValue } from "../../value"
-import { MotionPlugins } from "../../motion/context/MotionPluginContext"
 import { render } from "@testing-library/react"
 import { fireEvent } from "@testing-library/dom"
 import sync from "framesync"
 import { Constraints } from "../use-drag"
-
-type Point = {
-    x: number
-    y: number
-}
-
-const pos: Point = {
-    x: 0,
-    y: 0,
-}
-
-export const drag = (element: any) => {
-    pos.x = 0
-    pos.y = 0
-    fireEvent.mouseDown(element)
-
-    const controls = {
-        to: (x: number, y: number) => {
-            pos.x = x
-            pos.y = y
-            fireEvent.mouseMove(document.body, { buttons: 1 })
-
-            return controls
-        },
-        end: () => {
-            fireEvent.mouseUp(element)
-        },
-    }
-
-    return controls
-}
-
-export const MockDrag = ({ children }: { children: React.ReactNode }) => (
-    <MotionPlugins transformPagePoint={() => pos}>{children}</MotionPlugins>
-)
+import { MockDrag, drag } from "./utils"
 
 describe("dragging", () => {
     test("dragStart fires", async () => {