+++ src/gestures/types.ts
@@ -3,3 +3,5 @@ import { ValueAnimationControls } from "../motion"
 export interface ControlsProp {
     controls?: ValueAnimationControls
 }
+
+export type RemoveEvent = () => void