+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Framer Motion adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.5.1] Unreleased
 
+## Added
+
+-   `EventInfo` now passed as second argument to `onHoverStart` and `onHoverEnd`.
+
 ## Fixed
 
 -   Simplifying event system.